.github/actions/ci-status/requirements.txt
# Runtime dependencies for CI Status action
PyYAML==6.0.3
